fixed display of sparse dataframes

+++ pandas/sparse/array.py
@@ -244,7 +244,7 @@ to sparse
         """
         Dense values
         """
-        output = np.empty(len(self), dtype=np.float64)
+        output = np.empty(len(self), dtype=self.dtype)
         int_index = self.sp_index.to_int_index()
         output.fill(self.fill_value)
         output.put(int_index.indices, self)
